+++ CHANGELOG.md
@@ -8,6 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Added support for EOD close prices on Finage, IEX Cloud, Twelvedata, Tiingo and Unibit
+
+- New adapters:
+  - `expert-car-broker` to get car prices
+
 ### Changed
 
 ### Removed